Rearranged Columns import for speed
