+++ src/store.js
@@ -1,0 +1,9 @@
+// store.js
+import { reactive } from "vue";
+
+export const store = reactive({
+  isLoading: true,
+  loadingDone: () => {
+    isLoading: false;
+  },
+});
